refacotr: 함수 분리 및 OpenMarketManager extension 생성
