+++ OpenMarket/OpenMarket/Scene/ProductList/View/GridCollectionViewCell.swift
@@ -106,7 +106,7 @@ final class GridCollectionViewCell: UICollectionViewCell {
     }
     
     func setViewItems(_ product: ProductInformation) {
-        OpenMarketImageManager.setupImage(key: product.thumbnail) { image in
+        OpenMarketManager.setupImage(key: product.thumbnail) { image in
             DispatchQueue.main.async { [weak self] in
                 self?.productImageView.image = image
             }
