+++ OpenMarket/OpenMarket/Scene/ProductList/View/ListCollectionViewCell.swift
@@ -143,7 +143,7 @@ final class ListCollectionViewCell: UICollectionViewCell {
     }
     
     func setViewItems(_ product: ProductInformation) {
-        OpenMarketImageManager.setupImage(key: product.thumbnail) { image in
+        OpenMarketManager.setupImage(key: product.thumbnail) { image in
             DispatchQueue.main.async { [weak self] in
                 self?.productImageView.image = image
             }
